refactor(ui): add Dialog polishing, fixes, z-indices
