+++ packages/ui/src/design-tokens/spacing.data.ts
@@ -27,13 +27,13 @@ export const spacing = {
   $112: '112px',
   $116: '116px',
   $120: '120px',
-  $dialog_width_popup: '120px',
-  $dialog_width_browser: '480px',
   $148: '148px',
   $214: '214px',
   $294: '294px',
+  $312: '312px',
   $336: '336px',
   $342: '342px',
+  $480: '480px',
   $584: '584px',
   $fill: '100%',
   $auto: 'auto',
